Add files via upload
NAÏVE BAYES:
It is a supervised machine learning.
It is used for classification tasks.
It is based on bayes theorem for probability.
It is a probabilistic classifier, which means it predicts on the basis of the probability of an object
Why is it called Naïve Bayes?
The Naïve Bayes algorithm is comprised of two words Naïve and Bayes, Which can be described as:
o	Naïve: It is called Naïve because it assumes that the occurrence of a certain feature is independent of the occurrence of other features. Such as if the fruit is identified on the bases of color, shape, and taste, then red, spherical, and sweet fruit is recognized as an apple. Hence each feature individually contributes to identify that it is an apple without depending on each other.
o	Bayes: It is called Bayes because it depends on the principle of Bayes' Theorem.
Types of Naïve Bayes Model:
There are three types of Naive Bayes Model, which are given below:
o	Gaussian: The Gaussian model assumes that features follow a normal distribution. This means if predictors take continuous values instead of discrete, then the model assumes that these values are sampled from the Gaussian distribution.
o	Multinomial: The Multinomial Naïve Bayes classifier is used when the data is multinomial distributed. It is primarily used for document classification problems, it means a particular document belongs to which category such as Sports, Politics, education, etc.
The classifier uses the frequency of words for the predictors.
o	Bernoulli: The Bernoulli classifier works similar to the Multinomial classifier, but the predictor variables are the independent Booleans variables. Such as if a particular word is present or not in a document. This model is also famous for document classification tasks

Advantages of Naïve Bayes Classifier:
o	Naïve Bayes is one of the fast and easy ML algorithms to predict a class of datasets.
o	It can be used for Binary as well as Multi-class Classifications.
o	It performs well in Multi-class predictions as compared to the other Algorithms.
o	It is the most popular choice for text classification problems.
Disadvantages of Naïve Bayes Classifier:
o	Naive Bayes assumes that all features are independent or unrelated, so it cannot learn the relationship between features.
Applications of Naïve Bayes Classifier:
o	It is used for Credit Scoring.
o	It is used in medical data classification.
o	It can be used in real-time predictions because Naïve Bayes Classifier is an eager learner.
o	It is used in Text classification such as Spam filtering and Sentiment analysis.